Create `BlockHeader` when appending state diff
Motivation: block hashes are needed for comparisons between blocks into
Python DB vs Rust DB.
Block hashes are currently set as the block ids of the Python blocks.
