+++ crates/native_blockifier/src/py_state_diff.rs
@@ -6,7 +6,7 @@ use pyo3::prelude::*;
 use starknet_api::core::{ClassHash, ContractAddress, Nonce};
 use starknet_api::state::{StateDiff, StorageKey};
 
-use super::{NativeBlockifierError, NativeBlockifierResult};
+use crate::errors::{NativeBlockifierError, NativeBlockifierResult};
 use crate::py_utils::PyFelt;
 
 #[pyclass]
@@ -80,4 +80,12 @@ impl From<StateDiff> for PyStateDiff {
 
         Self { address_to_class_hash, address_to_nonce, storage_updates }
     }
+}
+
+#[derive(FromPyObject)]
+pub struct PyBlockInfo {
+    pub block_number: u64,
+    pub block_timestamp: u64,
+    pub gas_price: u128,
+    pub sequencer_address: PyFelt,
 }
